requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
mysql-connector==2.2.9
mysqlclient==2.2.4
psycopg2-binary==2.9.9
PyPDF2==3.0.1
sqlparse==0.4.4
